DNG writer: consolidate into imageio/imageio_dng.{c,h}
